Added ability to set the target image size
It's a bit awkward now because I couldn't figure out how to get dart to
send a u32 over ffi. The ffi.Uint32 class is non-constructable, and if I
put an ffi.Uint32 into the class signature, it threw an error saying two
identical signatures didn't match. So I guess it's just not ready yet.
Strings are fine for now I guess.